Conflicts:
	.idea/gradle.xml
	EzHerald/AndroidManifest.xml
	EzHerald/build.gradle
	EzHerald/res/layout/background_select_preference.xml
	EzHerald/res/layout/srtp_fragment_a.xml
	EzHerald/res/layout/srtp_fragment_b.xml
	EzHerald/res/values/array.xml
	EzHerald/res/values/strings.xml
	EzHerald/res/values/strings_activity_main_content_module_pref.xml
	EzHerald/src/com/herald/ezherald/BaseFrameActivity.java
	EzHerald/src/com/herald/ezherald/MainActivity.java
	EzHerald/src/com/herald/ezherald/api/APIAccount.java
	EzHerald/src/com/herald/ezherald/api/APIClient.java
	EzHerald/src/com/herald/ezherald/api/APIConf.java
	EzHerald/src/com/herald/ezherald/api/APIFactory.java
	EzHerald/src/com/herald/ezherald/api/APPID.java
	EzHerald/src/com/herald/ezherald/mainframe/MainContentFragment.java
	EzHerald/src/com/herald/ezherald/mainframe/MainMenuFragment.java
	EzHerald/src/com/herald/ezherald/srtp/FragmentA.java
	EzHerald/src/com/herald/ezherald/srtp/FragmentB.java
	EzHerald/src/com/herald/ezherald/srtp/Score.java
	EzHerald/src/com/herald/ezherald/srtp/SrtpActivity.java
	EzHerald/src/com/herald/ezherald/srtp/SrtpFragment.java
	EzHerald/src/com/herald/ezherald/srtp/SrtpGrabber.java
	EzHerald/src/main/jni/appid.cpp
	EzHerald/src/main/jni/empty.cpp

+++ EzHerald/src/com/herald/ezherald/srtp/SrtpGrabber.java
@@ -16,7 +16,7 @@ public class SrtpGrabber implements MainContentInfoGrabber {
     @Override
     public MainContentGridItemObj GrabInformationObject() {
         MainContentGridItemObj item = new MainContentGridItemObj();
-        double SrtpScore=score.getScore();
+        int SrtpScore=score.getScore();
         String updatetime=score.getUpdateTime();
         if(SrtpScore >= 0)
             item.setContent1(String.format("已获得Srtp学分",SrtpScore));
